python/pyarmnn/examples/image_classification/requirements.txt
requests>=2.23.0
urllib3>=1.25.8
Pillow>=6.1.0
numpy>=1.19.2
